Rename schema package to jsonschema
