+++ pkg/core/pipeline/condition/main.go
@@ -3,11 +3,11 @@ package condition
 import (
 	"fmt"
 
-	"github.com/invopop/jsonschema"
+	jschema "github.com/invopop/jsonschema"
+	"github.com/updatecli/updatecli/pkg/core/jsonschema"
 	"github.com/updatecli/updatecli/pkg/core/pipeline/resource"
 	"github.com/updatecli/updatecli/pkg/core/pipeline/scm"
 	"github.com/updatecli/updatecli/pkg/core/result"
-	"github.com/updatecli/updatecli/pkg/core/schema"
 )
 
 // Condition defines which condition needs to be met
@@ -92,11 +92,11 @@ func (c *Condition) Run(source string) (err error) {
 }
 
 // JSONSchema implements the json schema interface to generate the "condition" jsonschema.
-func (c Config) JSONSchema() *jsonschema.Schema {
+func (c Config) JSONSchema() *jschema.Schema {
 
 	type configAlias Config
 
 	anyOfSpec := resource.GetResourceMapping()
 
-	return schema.GenerateJsonSchema(configAlias{}, anyOfSpec)
+	return jsonschema.GenerateJsonSchema(configAlias{}, anyOfSpec)
 }
